fix issues: 1. fixed: view scroll when touch moving(rc-gesture has fixed this issue in v0.0.18) 2. fixed: sometimes buttons can not recover when two fingers touch moving
