+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rc-swipeout",
-  "version": "2.0.4",
+  "version": "2.0.5",
   "description": "swipe out ui component for react(web and react-native)",
   "keywords": [
     "react",
@@ -54,7 +54,7 @@
   "dependencies": {
     "babel-runtime": "6.x",
     "classnames": "2.x",
-    "rc-gesture": "^0.0.15",
+    "rc-gesture": "^0.0.18",
     "react-native-swipeout": "^2.2.2"
   },
   "devDependencies": {
